Actualizando la estructura de la base de datos
* Se actualizó el valor INT de todas las tablas por BIGINT

+++ SQL/inventory_schema.sql
@@ -8,27 +8,27 @@ CREATE DATABASE IF NOT EXISTS test_inventarios;
 USE test_inventarios;
 
 CREATE TABLE IF NOT EXISTS categories (
-	category_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	category_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	name VARCHAR(255) UNIQUE NOT NULL,
 	description VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS vendors (
-	vendor_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	vendor_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	name VARCHAR(255) UNIQUE NOT NULL,
 	description VARCHAR(255) NOT NULL,
 	address VARCHAR(255),
 	city VARCHAR(255),
 	state VARCHAR(255),
-	zip_code INTEGER,
-	phone_number INTEGER NOT NULL,
+	zip_code BIGINT,
+	phone_number BIGINT NOT NULL,
 	email_address VARCHAR(255),
 	website VARCHAR(255),
 	credit_limit FLOAT NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS products (
-	product_code INTEGER UNIQUE PRIMARY KEY NOT NULL,
+	product_code BIGINT UNIQUE PRIMARY KEY NOT NULL,
 	name VARCHAR(255) UNIQUE NOT NULL,
 	description VARCHAR(255) NOT NULL,
 	unit_measure VARCHAR(255) NOT NULL,
@@ -39,29 +39,29 @@ CREATE TABLE IF NOT EXISTS products (
 	demand FLOAT NOT NULL,
 	initial_quantity FLOAT NOT NULL,
 	final_quantity FLOAT,
-	category INTEGER NOT NULL,
-	vendor INTEGER NOT NULL,
+	category BIGINT NOT NULL,
+	vendor BIGINT NOT NULL,
 	FOREIGN KEY (category) REFERENCES categories(category_id),
 	FOREIGN KEY (vendor) REFERENCES vendors(vendor_id)
 );
 
 CREATE TABLE IF NOT EXISTS transactions (
-	transaction_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	transaction_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	name VARCHAR(255) NOT NULL,
 	description VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS finances (
-	finance_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	finance_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date DATE NOT NULL,
 	amount FLOAT NOT NULL,
 	description VARCHAR(255) NOT NULL,
-	transaction INTEGER NOT NULL,
+	transaction BIGINT NOT NULL,
 	FOREIGN KEY (transaction) REFERENCES transactions(transaction_id)
 );
 
 CREATE TABLE IF NOT EXISTS purchase_orders (
-	purchase_order_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	purchase_order_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
 	state VARCHAR(255) NOT NULL,
 	delivery_date DATE NOT NULL,
@@ -69,81 +69,81 @@ CREATE TABLE IF NOT EXISTS purchase_orders (
 );
 
 CREATE TABLE IF NOT EXISTS purchase_order_details (
-	purchase_order_details_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	purchase_order_details_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	quantity FLOAT NOT NULL,
 	price FLOAT NOT NULL,
-	product INTEGER NOT NULL,
-	purchase_order INTEGER NOT NULL,
+	product BIGINT NOT NULL,
+	purchase_order BIGINT NOT NULL,
 	FOREIGN KEY (product) REFERENCES products(product_code),
 	FOREIGN KEY (purchase_order) REFERENCES purchase_orders(purchase_order_id)
 );
 
 CREATE TABLE IF NOT EXISTS lots (
-	lot_number INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
-	purchase_order INTEGER NOT NULL,
+	lot_number BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	purchase_order BIGINT NOT NULL,
 	FOREIGN KEY (purchase_order) REFERENCES purchase_orders(purchase_order_id)
 );
 
 CREATE TABLE IF NOT EXISTS inventories (
-	inventory_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	inventory_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	quantity FLOAT NOT NULL,
 	unit_cost FLOAT NOT NULL,
 	purchase_date TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
 	expiration_date DATE NOT NULL,
-	product INTEGER NOT NULL,
-	purchase_order INTEGER NOT NULL,
-	lot INTEGER,
+	product BIGINT NOT NULL,
+	purchase_order BIGINT NOT NULL,
+	lot BIGINT,
 	FOREIGN KEY (product) REFERENCES products(product_code),
 	FOREIGN KEY (purchase_order) REFERENCES purchase_orders(purchase_order_id),
 	FOREIGN KEY (lot) REFERENCES lots(lot_number)
 );
 
 CREATE TABLE IF NOT EXISTS discounts (
-	discount_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	discount_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	discount_type VARCHAR(255) NOT NULL,
 	discount_amount FLOAT NOT NULL,
 	discount_start_date DATE NOT NULL,
 	discount_end_date DATE NOT NULL,
-	product INTEGER NOT NULL,
+	product BIGINT NOT NULL,
 	FOREIGN KEY (product) REFERENCES products(product_code)
 );
 
 CREATE TABLE IF NOT EXISTS payment_methods (
-	payment_method_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	payment_method_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	name VARCHAR(255) NOT NULL,
 	description VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS sales (
-	sale_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	sale_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date DATE NOT NULL,
 	total_amount FLOAT NOT NULL,
-	payment_method INTEGER NOT NULL,
+	payment_method BIGINT NOT NULL,
 	FOREIGN KEY (payment_method) REFERENCES payment_methods(payment_method_id)
 );
 
 CREATE TABLE IF NOT EXISTS sales_details (
-	sale_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	sale_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	quantity FLOAT NOT NULL,
 	price FLOAT NOT NULL,
-	product INTEGER NOT NULL,
+	product BIGINT NOT NULL,
 	FOREIGN KEY (product) REFERENCES products(product_code)
 );
 
 CREATE TABLE IF NOT EXISTS losses_type (
-	loss_type_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	loss_type_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	name VARCHAR(255) NOT NULL,
 	description VARCHAR(255) NOT NULL
 );
 
 CREATE TABLE IF NOT EXISTS losses (
-	losse_id INTEGER PRIMARY KEY NOT NULL AUTO_INCREMENT,
+	losse_id BIGINT PRIMARY KEY NOT NULL AUTO_INCREMENT,
 	date DATE NOT NULL,
 	description VARCHAR(255) NOT NULL,
 	quantity FLOAT NOT NULL,
 	amount FLOAT NOT NULL,
-	product INTEGER NOT NULL,
-	loss_type INTEGER NOT NULL,
+	product BIGINT NOT NULL,
+	loss_type BIGINT NOT NULL,
 	FOREIGN KEY (product) REFERENCES products(product_code),
 	FOREIGN KEY (loss_type) REFERENCES losses_type(loss_type_id)
 );
